feat: 메이커스 소개 레이아웃 쉬프트 수정
